fix: 결제 금액 타입 및 주문 금액 검증 보완

+++ src/main/java/org/example/groommvp/domain/payment/dto/PaymentResponse.java
@@ -11,7 +11,7 @@ import lombok.Builder;
 public record PaymentResponse(
 	Long paymentId,
 	Long orderId,
-	int amount,
+	Long amount,
 	PaymentStatus status,
 	LocalDateTime paidAt
 ) {
